feat(basket): counting badge for cart items
This commits adds a badge to the shopping
cart icon in the header that automatically
shows when there's at least one item in
the basket and vanishs when there are
zero items in the basket.

The commit also adds basic support for
sdk services to raise events that can
be listened on from within our app.