feat(thumbnailbar): added thumbnailbar directive
Created a thumbnailbar directive for the SDK.
This directive makes it easy to create a thumbnailbar
to be used on the product view.

It can be used like this:

    <cc-thumbnail-bar
        class="cc-thumbnail-bar-on-background"
        images="product.getAllImages()"
        on-change="onThumbnailSelected(image)">
    </cc-thumbnail-bar>

In order to achieve the directive the productService
had to be refactored.

All methods that previously lived on the productService
now live on the prototype of the product itself.

The productService was eventually killed.