feat(scrollingShadow): created new directive
The directive can be applied as an attribute
to any element that is scrollable.

Please note that the elements that create
the shadow effect need to be applied to the
parent element.

That comes with it's own limitations.
E.g. the parent element needs to be either
relative or absolute positioned.

Attemps that directly¬
placed them on the scrolled element, resulted¬
in poor performance (especially on mobile).¬

Fixes #30