feat(header): refactored fixed toolbar
The header was refactored to drop the
previous (sencha inspired) flexbox
approach were the scrolling happened in
a subdiv. Instead we now use the much
simpler "position:fixed" approach if
we know that the browser is capable to
handle it gracefully. If not, the toolbar
just scrolls away with the content.

Unfortunately there is no easy way to tell
if the browser supports fixed positioning
properly. So for that matter we rely
on UA sniffing.

It's actually the recommended practice by
Brad Frost:
http://bradfrostweb.com/blog/mobile/fixed-position/

By now we only use the fixed toolbar for
Android > 2.2 and iOS > 4.

We also removed the overthrow.js library since
we don't need it anymore as a polyfill.

Unfortunately this commit breaks the shadow
scroller implementation.

Fixes #52, #33, #41

Reopens #30