chore(grunt): enhanced grunt tasks
With this commit all templates are precompiled
into js modules so that there is no need
anymore to load templates via XHR.

BREAKING:
Another thing that has been changed is that
the app index.html file is no longer the file
to be used to view the app. The file to
kick off the app is now dist/index.html.

This is so that one no longer has to manually
update the index.html file

Fixex #5