feat(checkout): provide basic checkout methods
This commit brings the basic checkout methods (read:
everything we support except paypal) to the app.

It also adds a bunch of new things to the SDK
that were developed to assist app developers with
the checkout.

It also brings tests and a bunch of new mocking
capabilities.

It's a huge commit and some unrelated changes
slipped through.

Please note that the checkout is still half-assed
as validations and error handling are still missing.
Please refer to the ticket to track progress:

https://github.com/couchcommerce/frontend-spike/issues/65