fix(checkout): preserve payment and shipping method
The payment method and the shipping method
are now preserved within the lifetime of the
app. In contrast to the addresses it's not
persisted via the storageService as
it probably does not make sense to present the
use the last used checkout methods on his
next shopping tour since those can change
based on the bought items.

This feature is most important for going back
from the summary page to the checkout page.

It also fixes a potential issue with the
select box when the underlying data source
changes and removes the currently selected
item.