feat($dialog): allow usage of html messages
We monkeypatch the bootstrap $dialog service
to use a different template that uses
html-bind-unsafe instead of a regular
bind. This allows us to use it to show
cms pages as overlays.

It's also great to further allow us to restyle
the dialog without making any modifications
to the original files.