chore(angular): update to a more recent angular build
This commit sets angular to a custom build (as before)
that is more recent.

It's this one: 11521a4cde689c2bd6aaa227b1f45cb3fb53725b

It's the commit before matsko started rewriting
ngAnimate to be entirely class based. The class
based ngAnimate is just not ready yet and afaik
it doesn't yet work together with the ui-router.

We will upgrade to a more recent build at a
later stage once the dust around ngAnimate has
settled.