fix(*): prevent horizontal scrollbars on iPhone 4
On iPhone with iOS <=6 there seems to be a bug
with absolute positioned elements causing
horizontal scrollbars for no reason.

We need to position the pages absolutely during
the page transitions. Previously they were
just absolutely positioned all the time, now
we just absolutely position them during the
transition and have them positioned relative
for the rest of the time