fix(basketService): cumulate same products across app reloads
Due to a bug in cc.Util.deepExtend "null" values
were omitted when extending two objects. That
in turn caused the basketService to think he
deals with two different products even so it
were the same.

We flipped our cc.Util.deepExtend against a
simpler cc.Util.extend(). If one needs to have
a clone there is cc.Util.clone() which does
exactly that.