fix(css): fix ugly 3d effect on android 2.3
For our buttons and cards we previously
used a bottom-border-radius of 3px to create
a 3d like effect.

However, that doesn't work well for the
native Android browser (not Chrome for Android).
The browser renders it poorly. To fix that we
can achieve the same effect with box-shadow
which looks much nicer.