fix(scrolling): remove cc-scroller--if-no-fixed
It was never needed to set those styles for
when no position fixed is supported. Furthermore
it breaks document.body.scrollTop in a way
that it does not report/accept scrollOffset
anymore.

This commit was triple checked against all
previous cases and does not seem to break
anything for any supported devices.