feat(scrollPositionService): reset scrollPosition only in cases that make sense
Previously we always reset the scrollPosition to
the position the view had the last time it was
displayed.

That doesn't make sense for every case. Actually,
most of the time we want the views to be set to
the very top. However, if we go back from
product > products > categories then we
*DO* want the view to be reset to the last known
scroll position because that shows the product/category
where the user just came from.