fix(configService): use configService to retrieve apiHttpMethod
The apiHttpMethod is a configuration that was
introduced mainly to be used for development
during offline phases such as train rides etc.

It was not meant to be exposed to the customer.
Therefore we should retrieve it through the
configService and set the default value
to jsonp. This means, if it's missing from
the config file (which is the case for hosted shops)
it will just default to jsonp. However, during
development one can just set it to 'get' if needed.