chore(app): move base shell out of index.html
We want to keep the index.html as slim as possible.
That's mostly because the index will be generated by the
backend and therfore all changes to this file need to be
communicated and synced with the backend.

We were now able to move all the basic shell structure
into it's own shell.tpl.html file