fix(app): repair deep linking of products
The deep linking broke when we shifted
code from index.html into a shell.tpl.html
and put a ng-include on the body. There seems
to be a bug in angular with ng-include
on the body breaking jsonp requests.

We work around that by using cc-include
in a subdiv to achieve the same effect.

Fixes #139