fix(stateChangeService): always detect parent/child relation correctly
Previously relationships between categories were
not detected correctly when they were not in a
*direct* relationship.

Simply speaking previously a was detected as parent
of b but not as parent of c. This caused a bug
where the wrong page transition was used when one
jumped between categories that were not directly
next to each other.

Fixes #147