feat(paypal): handle paypal checkout from checkout view
This commit makes it so that it brings up the PayPal
overlay from within the checkout view if one chooses
PayPal as payment method.

In the old app the app transitioned back to the
cart view if one had choosen PayPal from within the
checkout view. However, that doesn't seem to be the
logical choice anymore. Showing the overlay right
where we are seems cleaner.

Fixes #154