refactor(cc-select-box): decouple from couch theming
The component is now fully decoupled from the
CouchCommerce theming and is therfore one step
closer to be released as a standalone component.

What this means:

1. Nothing in the sdk/directives/ccSelectBox directory
knows about things like $content-background-color,
$background-color etc.

2. The component defines their own variables to be
exposed

3. The component creates it's classes through a mixin
which gives it the chance to stamp out lots of variations
like cc-select-box--dark, cc-selectbox--light to be used
in parallel within one app.

4. The component is configured to match the CouchCommerce
theming through a dedicated cc-select-box-couch-config.scss file
which sits in the app (!!!).

Relates to #180