fix(checkoutService): glue redirect to token
When the previously saved redirect is used
inside the activateOrder(token) function,
we should make sure that it actually was
intended to be used together with the
current token.

Otherwise we might get into trouble if
we also add payment methods where getSummary()
won't be called (e.g. because the summary
is displayed directly at a 3rd party page).

It could happen that inside activateOrder
we still have an old redirect saved which ends
up being used accidentially.