fix(category): prevent duplicate transitions
Previously the CategoryController redirected
to the product list if it detected that the
category does not contain any subcategories.
However, when the CategoryController runs, it
basically means that it's transition was also
used. Means: in such cases the user saw to
transitions happening which feels odd.

By moving this check into the routing code
we can intercept *before* any transition runs
and then redirect to the proper route.

Relates to #64