feat(checkout): introduces info tooltip in checkout process
This commit adds a simple css class called `cc-tooltip` which
act as a base class and also a corresponding modifier `cc-tooltip--info`.

This makes sure that once we need a tooltip for warnings or errors
we can simply extend the existing base class with a corresponding
modifier.

The tooltip is currently only used in the checkout process and
appears if the checkout form isn't filled out correctly.

I also quick-added a new translation id to fill the tooltip with
proper text.

Fixes #224