chore(grunt): introduces new build process
Here it is! Our new build process. This PR does the following:

- introduces new tasks build, compile, compile-debug, deploy, deploy-debug
- removes the unsed stable angular version
- updates the .jshintrc
- adds build.config.js
- introduces a vendor folder (here's our custom angular version installed)
- removes lib folder
- introduces assets folder
- Our new build process is splitted into two tasks build and compile.
- moves files from `modules` to `src/app`
- moves scss files under `assets` folder
- makes grunt also watching vendor js files
- moves grunt-conventional-changelog to `tasks` folder

The build task generates a running app. This is for local development. It also
starts a server for you so you don't have to care bout that :)

The compile task takes the generated stuff from the build task and gives you a
production ready app.