refactor(ccImageZoom): extracted some parts into a new home
This introduces two new things:

ccImageZoomDomActors:

This is basically just an object that holds
all the different DOM nodes which take part
in the process. This is a base object other
services can depend on to manipulate such
DOM elements.

ccImageZoomMaskService:

This just deals with everything around the
mask.

Relates to #259