refactor(ccImageZoom): refactored `inAnimation` var  out of lerpToPosition
the `inAnimation` variable doesn't seem to have a
business inside the `lerpToPosition` method. The only
important thing it seems is to set it to `true` once
the lerp starts and set it to `false` once it's over.
This should be done from outside hence the new `lerp`
method which simply wraps the `lerpToPosition` method.

This is another crucial step to get `lerpToPosition`
side effect free.