refactor(ccImageZoom): extract lerpAnim service
Finally moved the whole `lerpToPosition` method into
it's own service. This could be even moved into a
generic anim module in the sdk later as it doesn't have
any coupling to the image zoom any more. Could also
be used for animations where we want to animate an
item flying into the basket.