fix(cart): improve perf/reliability of plus/minus buttons
Turns out basketService.getSummary() is quite costly.
On some devices (iPhone, I'm looking at you) there's not enough
time left in our budget to do the calculation within the given frame.
This causes quite weird behaviour (dropped frames?) where the screen
isn't updated and then when the user tappes the next time the amount of the
item jumps two steps (or better that's how it looks for the user, in reality
most likely it was just the screen which wasn't updated correctly before)
We overcome this by scheduling the getSummary() for the next frame.