feat(grunt): improves deployment process
This introduces a couple of new commands:

grunt stack-syms
grunt pop-syms

It also makes sure that grunt deploy and
it's sisters (deploy-debug, deploy-dry-run)
by default reinstall all npm packages.

This, however, can be changed with the
--force-local flag which makes sure that
exactly what runs on the developers local
machine will be deployed as is.

Fixed #463